% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{aic}
\alias{aic}
\title{Akaike Information Criterion (AIC)}
\usage{
aic(x, sequences = NULL)
}
\arguments{
\item{x}{An object for which there exists a \code{loglik} attribute if
\code{sequences = NULL} or a \code{loglik} method otherwise.}

\item{sequences}{Optional. A list of vectors representing the sequences for
which the AIC will be computed based on \code{x} using the method \code{loglik}.}
}
\value{
Value of the AIC.
}
\description{
Generic function computing the Akaike Information Criterion of
the model \code{x}, with the list of sequences \code{sequences}.
}
