% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_median.R
\name{ci_median}
\alias{ci_median}
\title{Confidence interval for median by inverting sign test}
\usage{
ci_median(x, conf.level = 0.95)
}
\arguments{
\item{x}{vector of data}

\item{conf.level}{level for CI (as decimal), default 95 percent}
}
\value{
lower and upper limits
}
\description{
Confidence interval for median by inverting sign test
}
\examples{
ci_median(mtcars$mpg,0.90)
ci_median(1:20)
ci_median(1:5)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
