% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_median.R
\name{ci_median0}
\alias{ci_median0}
\title{Confidence interval for median by inverting sign test}
\usage{
ci_median0(x, conf.level = 0.95, tol = 0.01)
}
\arguments{
\item{x}{vector of data}

\item{conf.level}{level for CI (as decimal), default 95 percent}
}
\value{
lower and upper limits
}
\description{
Confidence interval for median by inverting sign test
}
\examples{
ci_median0(mtcars$mpg)
ci_median0(mtcars$mpg,0.80)
ci_median0(1:20)
ci_median0(1:5)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
