% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{convolution}
\alias{convolution}
\title{Discrete-time convolution product of \eqn{f} and \eqn{g}
(See definition 2.2 p. 20)}
\usage{
convolution(f, g)
}
\arguments{
\item{f}{A vector giving the values of the function \eqn{f} for each
\eqn{k \in \mathbb{N}}.}

\item{g}{A vector giving the values of the function \eqn{g} for each
\eqn{k \in \mathbb{N}}.}
}
\value{
A vector giving the values of the discrete-time convolution of
\eqn{f} and \eqn{g} for each \eqn{k \in \mathbb{N}}.
}
\description{
Discrete-time convolution product of \eqn{f} and \eqn{g}
(See definition 2.2 p. 20)
}
