% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{get.f}
\alias{get.f}
\title{Method to get the conditional sojourn time distribution f}
\usage{
get.f(x, k)
}
\arguments{
\item{x}{An object of S3 class \code{smmfit} or \code{smm}.}

\item{k}{A strictly positive integer giving the time horizon (k must be
different from 0 since we assume that there are not instantaneous
transitions).}
}
\value{
A vector, matrix or array giving the value of \eqn{f} at each time
between 1 and \code{k}.
}
\description{
Computes the conditional sojourn time distribution \eqn{f(k)},
\eqn{f_{i}(k)}, \eqn{f_{j}(k)} or \eqn{f_{ij}(k)}.
}
