% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.H.R
\name{get.H}
\alias{get.H}
\title{Function to compute the value of the sojourn time cumulative distribution \eqn{H}}
\usage{
get.H(q)
}
\arguments{
\item{q}{An array giving the values of the kernel for a giving time horizon
\eqn{[0, \dots, k]} (This kernel \code{q} is the output of the method \code{getKernel}
or \code{get.qy}).}
}
\value{
An array giving the value of \eqn{H(k)} at each time between 0
and \code{k}.
}
\description{
Function to compute the value of \eqn{H} (See equation (3.4) p.46).
}
