% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{get.Kpar}
\alias{get.Kpar}
\title{Method to get the number of parameters of a Markov or semi-Markov chain}
\usage{
get.Kpar(x)
}
\arguments{
\item{x}{An object for which the number of parameters can be returned (An
object of class \code{smm} or \code{mm}).}
}
\value{
A positive integer giving the number of parameters.
}
\description{
Method to get the number of parameters of a Markov or
semi-Markov chain. This method is useful for the computation of criteria
such as AIC and BIC.
}
