% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{get.limitDistribution}
\alias{get.limitDistribution}
\title{Method to get the limit (stationary) distribution}
\usage{
get.limitDistribution(x, klim = 10000)
}
\arguments{
\item{x}{An object of S3 class \code{smmfit} or \code{smm}.}

\item{klim}{Optional. The time horizon used to approximate the series in the
computation of the mean sojourn times vector \eqn{m} (cf.
\link{meanSojournTimes} function).}
}
\value{
A vector of length \eqn{\textrm{card}(E)} giving the values of the
limit distribution.
}
\description{
Method to get the limit (stationary) distribution of a
semi-Markov chain.
}
