% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Niuj.R
\name{get.Niuj}
\alias{get.Niuj}
\title{Function giving the value of the counting process Niuj used in the
estimation of the kernel and the transition matrix of censored and
non-parametric semi-markov chains (cf. article Exact MLE and asymptotic
properties for nonparametric semi-Markov models)}
\usage{
get.Niuj(sequences)
}
\arguments{
\item{sequences}{A list of sequences of states.}
}
\value{
An array giving the values of the counting process \eqn{N_{iuj}}.
}
\description{
Function giving the value of the counting process Niuj used in the
estimation of the kernel and the transition matrix of censored and
non-parametric semi-markov chains (cf. article Exact MLE and asymptotic
properties for nonparametric semi-Markov models)
}
