% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.psi.R
\name{get.psi}
\alias{get.psi}
\title{Function to compute the value of the matrix-valued function \eqn{\psi}}
\usage{
get.psi(q)
}
\arguments{
\item{q}{An array giving the values of the kernel for a giving time horizon
\eqn{[0, \dots, k]} (This kernel \code{q} is the output of the method \code{getKernel}
or \code{.get.qy}).}
}
\value{
An array giving the value of \eqn{\psi(k)} at each time between 0
and \code{k}.
}
\description{
Function to compute the value of \eqn{\psi}, the matrix-valued
function (See equation (3.16) p.53).
}
