% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{get.Py}
\alias{get.Py}
\title{Method to compute the value of \eqn{P_{Y}}}
\usage{
get.Py(x, k, upstates = x$states)
}
\arguments{
\item{x}{An object of class \code{smm}.}

\item{k}{A positive integer giving the time horizon.}

\item{upstates}{Vector giving the subset of operational states \eqn{U}.}
}
\value{
An array giving the value of \eqn{P_{Y}(k)} at each time between 0
and \code{k}.
}
\description{
Method to compute the value of \eqn{P_{Y}}
(See Proposition 5.1 p.105-106).
}
