% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{get.stationaryDistribution}
\alias{get.stationaryDistribution}
\title{Method to get the stationary distribution}
\usage{
get.stationaryDistribution(x)
}
\arguments{
\item{x}{An object of S3 class \code{mm} or \code{mmfit}.}
}
\value{
A vector of length \eqn{\textrm{card}(E)} giving the values of the
stationary distribution.
}
\description{
Method to get the stationary distribution of a Markov chain.
If the order of the Markov chain is higher than one, a block matrix is
computed to get the stationary distribution.
}
