% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProcesses.R
\name{getProcesses}
\alias{getProcesses}
\title{Function to compute processes based on a list of sequences}
\usage{
getProcesses(sequences)
}
\arguments{
\item{sequences}{A list of sequences of states.}
}
\value{
A list giving:
\itemize{
\item{s : }{Cardinal of the states space;}
\item{states : }{Vector of state space of length \eqn{s};}
\item{M : }{The total length (sum of the length of each sequence);}
\item{L : }{The number of sequences;}
\item{kmax : }{Maximum length of the sojourn times;}
\item{Ym: }{List of semi-Markov chains;}
\item{Jm: }{List of successively visited states for each sequence;}
\item{Tm: }{List of successive time points when state changes for each sequence;}
\item{Lm: }{List of sojourn time for each sequence;}
\item{Um: }{List of backward recurrence time for each sequence;}
\item{counting : }{List giving the counting processes such as
\eqn{N_{ij}(k)}, \eqn{N_{i}(k)}, \eqn{N_{j}(k)}...}
}
}
\description{
Function to compute processes based on a list of sequences
}
