% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmfit.R
\name{is.smmfit}
\alias{is.smmfit}
\title{Function to check if an object is of class \code{smmfit}}
\usage{
is.smmfit(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
\code{is.smmfit} returns \code{TRUE} or \code{FALSE} depending on whether \code{x} is an
object of class \code{smmfit} or not.
}
\description{
\code{is.smmfit} returns \code{TRUE} if \code{x} is an object of class \code{smmfit}.
}
