% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmnonparametric.R
\name{is.smmnonparametric}
\alias{is.smmnonparametric}
\title{Function to check if an object is of class \code{smmnonparametric}}
\usage{
is.smmnonparametric(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
\code{is.smmnonparametric} returns \code{TRUE} or \code{FALSE} depending on whether
\code{x} is an object of class \code{smmnonparametric} or not.
}
\description{
\code{is.smmnonparametric} returns \code{TRUE} if \code{x} is an object of
class \code{smmnonparametric}.
}
