% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smmparametric.R
\name{is.smmparametric}
\alias{is.smmparametric}
\title{Function to check if an object is of class \code{smmparametric}}
\usage{
is.smmparametric(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
\code{is.smmparametric} returns \code{TRUE} or \code{FALSE} depending on whether
\code{x} is an object of class \code{smmparametric} or not.
}
\description{
\code{is.smmparametric} returns \code{TRUE} if \code{x} is an object of
class \code{smmparametric}.
}
