% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allgenerics.R
\name{loglik}
\alias{loglik}
\title{Log-likelihood Function}
\usage{
loglik(x, sequences = NULL)
}
\arguments{
\item{x}{An object for which there exists a \code{loglik} attribute if
\code{sequences = NULL}. Otherwise, the log-likelihood will be computed using
the model \code{x} and the sequences \code{sequences}.}

\item{sequences}{Optional. A list of vectors representing the sequences for
which the log-likelihood will be computed based on \code{x}.}
}
\value{
Value of the log-likelihood.
}
\description{
Generic function computing the log-likelihood of the model \code{x},
with the list of sequences \code{sequences}.
}
