% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixConvolution.R
\name{matrixConvolution}
\alias{matrixConvolution}
\title{Discrete-time matrix convolution product
(See definition 3.5 p. 48)}
\usage{
matrixConvolution(a, b)
}
\arguments{
\item{a}{An array of dimension \eqn{(S, S, k + 1)}.}

\item{b}{An array of dimension \eqn{(S, S, k + 1)}.}
}
\value{
An array of dimension \eqn{(S, S, k + 1)} giving the discrete-time
matrix convolution product for each \eqn{k \in \mathbb{N}}.
}
\description{
Discrete-time matrix convolution product
(See definition 3.5 p. 48)
}
