% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_test.R
\name{median_test}
\alias{median_test}
\title{Mood's median test for comparison of group medians}
\usage{
median_test(d, x, g, tol = 1e-06)
}
\arguments{
\item{d}{a data frame}

\item{x}{unquoted name of quantitative variable}

\item{g}{unquoted name of grouping variable}

\item{tol}{(default 1e-6) any data values closer to overall median than this are discarded}
}
\value{
list of 2 objects: table, counts of values above and below the grand median in each group; value, test statistic, df and P-value
}
\description{
Mood's median test for comparison of group medians
}
\examples{
median_test(datasets::mtcars, mpg, cyl)
d <- data.frame(z = 1:9, gg = c(1, 1, 1, 1, 1, 2, 2, 2, 2))
median_test(d, z, gg)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
