% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_test.R
\name{median_test0}
\alias{median_test0}
\title{Mood's median test for comparison of group medians}
\usage{
median_test0(x, g)
}
\arguments{
\item{x}{vector of data}

\item{g}{vector of group memberships (same length as x)}
}
\value{
list of 2 objects: table, counts of values above and below the grand median in each group; value, test statistic, df and P-value
}
\description{
Mood's median test for comparison of group medians
}
\examples{
median_test0(mtcars$mpg,mtcars$cyl)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
