% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_median_test.R
\name{pairwise_median_test}
\alias{pairwise_median_test}
\title{Pairwise Mood's median tests for all comparison of pairs of group medians}
\usage{
pairwise_median_test(d, x, g, tol = 1e-06)
}
\arguments{
\item{d}{a data frame}

\item{x}{unquoted name of quantitative variable}

\item{g}{unquoted name of grouping variable (can be a factor, is treated as text)}

\item{tol}{(default 1e-6) any data values closer to overall median than this are discarded}
}
\value{
data frame of groups being compared and unadjusted and Bonferroni-adjusted P-values
}
\description{
Pairwise Mood's median tests for all comparison of pairs of group medians
}
\examples{
pairwise_median_test(mtcars,mpg,cyl)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
