% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_sign.R
\name{pval_sign0}
\alias{pval_sign0}
\title{Two-sided P-value for sign test}
\usage{
pval_sign0(med0, x)
}
\arguments{
\item{med0}{null median}

\item{x}{vector of data for test}
}
\value{
P-value of two-sided sign test for median
}
\description{
Two-sided P-value for sign test
}
\examples{
pval_sign0(3.5, 1:10)
pval_sign0(3, 1:10)
pval_sign0(25, datasets::mtcars$mpg)
}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
