% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setSeed}
\alias{setSeed}
\title{Set the RNG Seed from within Rcpp}
\usage{
setSeed(seed)
}
\arguments{
\item{seed}{An \code{unsigned int} that is the seed one wishes to use.}
}
\value{
A set RNG scope.
}
\description{
Set the RNG Seed from within Rcpp
}
\examples{
set.seed(10)
x <- rnorm(5, 0, 1)
setSeed(10)
y <- rnorm(5, 0, 1)
all.equal(x, y, check.attributes = FALSE)

}
