% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_test.R
\name{sign_test}
\alias{sign_test}
\title{Sign test for given median}
\usage{
sign_test(med0, x)
}
\arguments{
\item{med0}{null median}

\item{x}{vector of data for test}
}
\value{
list of two elements: table of values above and below null median, data frame of 1-sided and 2-sided P-values
}
\description{
Sign test for given median
}
\examples{
sign_test(3.5,1:10)
sign_test(3,1:10)
sign_test(25,mtcars$mpg)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
