% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_test0.R
\name{sign_test0}
\alias{sign_test0}
\title{Sign test (simplified) for given median}
\usage{
sign_test0(x, med0 = 0, tol = 1e-06)
}
\arguments{
\item{x}{vector of data}

\item{med0}{null median (defaults to zero)}

\item{tol}{(default 1e-6) how close a data value has to be to the null median to be considered equal to null median (and discarded)}
}
\value{
list of two elements: table of values above and below null median, data frame of 1-sided and 2-sided P-values
}
\description{
Sign test (simplified) for given median
}
\examples{
sign_test0(1:10,3.5)
sign_test0(1:10,3)
sign_test0(datasets::mtcars$mpg,25)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
