% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_median.R
\name{ci_median}
\alias{ci_median}
\title{Confidence interval for median by inverting sign test}
\usage{
ci_median(d, x, conf.level = 0.95, tol = 0.01)
}
\arguments{
\item{d}{a data frame}

\item{x}{unquoted name of column of data}

\item{conf.level}{level for CI (as decimal), default 95 percent}

\item{tol}{ends of CI determined to within this accuracy, default 0.01}
}
\value{
lower and upper limits
}
\description{
Confidence interval for median by inverting sign test
}
\examples{
d=data.frame(z=1:20)
d1=data.frame(z=1:5)
ci_median(d,z)
ci_median(d1,z)

}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
