% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmfit.R
\name{is.mmfit}
\alias{is.mmfit}
\title{Function to check if an object is of class \code{mmfit}}
\usage{
is.mmfit(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
\code{is.mmfit} returns \code{TRUE} or \code{FALSE} depending on whether \code{x} is an
object of class \code{mmfit} or not.
}
\description{
\code{is.mmfit} returns \code{TRUE} if \code{x} is an object of
class \code{mmfit}.
}
