% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm.R
\name{is.smm}
\alias{is.smm}
\title{Function to check if an object is of class \code{smm}}
\usage{
is.smm(x)
}
\arguments{
\item{x}{An arbitrary R object.}
}
\value{
\code{is.smm} returns \code{TRUE} or \code{FALSE} depending on whether \code{x} is an
object of class \code{smm} or not.
}
\description{
\code{is.smm} returns \code{TRUE} if \code{x} is an object of class \code{smm}.
}
