% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Markov model specification}
\usage{
mm(states, init, ptrans, k = 1)
}
\arguments{
\item{states}{Vector of state space of length s.}

\item{init}{Vector of initial distribution of length s ^ k.}

\item{ptrans}{Matrix of transition probabilities of dimension \eqn{(s, s)}.}

\item{k}{Order of the Markov chain.}
}
\value{
An object of class \link{mm}.
}
\description{
Creates a model specification of a Markov model.
}
\seealso{
\link{simulate.mm}, \link{fitmm}
}
