% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_test.R
\name{sign_test}
\alias{sign_test}
\title{Sign test for given median}
\usage{
sign_test(d, x, med0 = 0, tol = 1e-06)
}
\arguments{
\item{d}{a data frame}

\item{x}{unquoted name of column to test}

\item{med0}{null median (defaults to zero)}

\item{tol}{(default 1e-6) how close a data value has to be to the null median to be considered equal to null median (and discarded)}
}
\value{
list of two elements: table of values above and below null median, data frame of 1-sided and 2-sided P-values
}
\description{
Sign test for given median
}
\examples{
d=data.frame(z=1:10)
sign_test(d,z,3.5)
sign_test(d,z,3)
sign_test(mtcars,mpg,25)
}
\author{
Ken Butler, \email{butler@utsc.utoronto.ca}
}
